finalizing locations of tests
